@@ -15,7 +15,7 @@
     "start": "node server.js"
   },
   "dependencies": {
-    "react": "^0.12",
+    "react": "git+https://github.com/facebook/react#5f4d759a",
     "react-router": "^0.11.4",
     "react-style": "git+https://github.com/js-next/react-style#inline-styles",
     "rest": "git+https://github.com/cujojs/rest#f40242a43a3db826af",
