@@ -2,7 +2,6 @@ var React = require('react');
 var Invariant = require('react/lib/invariant');
 var UI = require('ui/index');
 var StyleKeys = require('ui/lib/StyleKeys');
-var AnimateStore = require('ui/stores/AnimateStore');
 
 // This is a somewhat haphazard mixin at the moment that
 // organizes and runs animations on elements.
@@ -43,104 +42,39 @@ var AnimateStore = require('ui/stores/AnimateStore');
 // - animationActive: for disabling animations
 
 var defined = variable => (typeof variable !== 'undefined');
+var pick = (a, b) => typeof a !== 'undefined' ? a : b;
 
 module.exports = {
   contextTypes: {
-    animationProps: React.PropTypes.object,
-    animationsActive: React.PropTypes.bool
+    animations: React.PropTypes.object,
+    animationsDisabled: React.PropTypes.bool
   },
 
-  componentWillUpdate(nextProps, nextState) {
-    this._setAnimationsState(nextProps, nextState);
-  },
-
-  // This is what links it to the UI loading, and grabs animations
-  // todo: make this not necessary
   getAnimator(animation) {
     return UI.getAnimations(animation);
   },
 
-  // Helper props
   animationsDisabled() {
     return this.props.animationDisabled || this.context.animationDisabled;
   },
 
-  isAnimating(source) {
-    return this._animationsState && this.getAnimationState(source).step % 1 !== 0;
+  isAnimating(opts) {
+    return this.getAnimation(opts).step % 1 !== 0;
   },
 
-  getAnimationProp(animation, animations) {
-    animations = animations || this.getAnimations();
-    return animations.length && animations.filter(a => a && a.animation === animation);
-  },
-
-  hasAnimation(animation, animations) {
+  hasAnimation() {
     animations = animations || this.getAnimations();
     return animations.length && !!this.getAnimationProp(animation, animations).length;
   },
 
-  getAnimations(props, state) {
-    var animations = [];
-
-    if (typeof props === 'undefined') {
-      props = this.props;
-      state = this.state;
-    }
-
-    if (props.animations && props.animations.length)
-      animations = props.animations;
-
-    if (state && state.animations && state.animations.length)
-      animations = animations.concat(state.animations);
-
-    return animations;
-  },
-
-  _setAnimationsState(props, state) {
-    if (typeof props === 'undefined') {
-      props = this.props;
-      state = this.state;
-    }
-
-    this._animationsState = {};
-    var animations = this.getAnimations(props, state);
-
-    if (animations && animations.length) {
-      var source;
-
-      animations.forEach(animation => {
-        source = animation.source || 'self';
-
-        this._animationsState[source] = Object.assign({},
-          AnimateStore(animation.source),
-          props && props.animationProps && props.animationProps[animation.source],
-          this.context && this.context.animationProps && this.context.animationProps[animation.source],
-          source === 'self' && state && defined(state.step) && { step: state.step },
-          source === 'self' && state && defined(state.index) && { index: state.index }
-        );
-      });
-    }
-  },
-
   getAnimationState(source) {
-    if (!this._animationsState)
-      this._setAnimationsState();
 
-    return this._animationsState[source || 'self'];
   },
 
-  getAnimationStep(source) {
-    if (!this._animationsState)
-      this._setAnimationsState();
-
-    return this.getAnimationState(source).step;
-  },
-
-  // this is the meat of it, fetches animations and returns an object
-  // of animations. key is the ref, value is an object of styles.
-  // TODO: make this use RAF and return on every frame
-  getAnimation(name) {
-    var animations = this.getAnimations();
+  getAnimationStyle({ ref, source }) {
+    source = source || this._animationSource || 'self';
+    var animations = this.getAnimations(ref);
+    var animationState = this.getAnimationState(ref);
     var animation, i, len = animations.length;
 
     if (!len)
@@ -210,80 +144,5 @@ module.exports = {
       transformString += `translate3d(${t.translate.x || 0}px, ${t.translate.y || 0}px, ${t.translate.z || 0}px)`;
 
     return transformString;
-  },
-
-  // experimental: for out of react render
-  // _animate() {
-  //   if (!this.animationsDisabled() && this.getAnimations().length && !this.hasPendingAnimations) {
-  //     animationQueue.push(this);
-  //     this.hasPendingAnimations = true;
-  //     window.requestAnimationFrame(runAnimations);
-  //   }
-  // },
-
-  // // experimental: for out of react render
-  // _setAnimationSStatetyles() {
-  //   this.hasPendingAnimations = false;
-  //   var animations = this.getAnimation();
-
-  //   if (!animations || this._lifeCycleState !== 'MOUNTED')
-  //     return;
-
-  //   var node, selector;
-  //   var hasHeadStyleTag = !!this._headStyleTag;
-
-  //   if (!hasHeadStyleTag) {
-  //     this._headStyleTag = document.createElement('style');
-  //     this._headStyleTag.ref = this._uniqueID;
-  //   }
-  //   else {
-  //     this._headStyleTag.innerHTML = '';
-  //   }
-
-  //   Object.keys(animations).forEach(key => {
-  //     node = key === 'self' ?
-  //       this.getDOMNode() :
-  //       this.refs[key] && this.refs[key].getDOMNode();
-
-  //     if (!node)
-  //       return;
-
-  //     selector = node.id ?
-  //       `#${node.id}` :
-  //       `[data-reactid="${node.getAttribute('data-reactid')}"]`;
-
-  //     // set tag styles
-  //     this._headStyleTag.innerHTML +=
-  //       `${selector} {
-  //         ${this._stylesToString(animations[key])}
-  //       }`;
-  //   });
-
-  //   if (!hasHeadStyleTag)
-  //     document.head.appendChild(this._headStyleTag);
-  // },
-
-  // // experimental: for out of react render
-  // _stylesToString(obj) {
-  //   return Object.keys(obj).reduce(
-  //     (acc, key) =>
-  //       `${key}: ${obj[key]};
-  //        ${acc}`,
-  //     '');
-  // },
-
-  // componentWillUnmount() {
-  //   if (this._headStyleTag)
-  //     document.head.removeChild(this._headStyleTag);
-  // },
+  }
 };
-
-// experimental: out of react animations
-// var animationQueue = [];
-// function runAnimations() {
-//   var i, len = animationQueue.length;
-//   for (i = 0; i < len; i++) {
-//     animationQueue[i]._setAnimationSStatetyles.call(animationQueue[i]);
-//   }
-//   animationQueue = [];
-// }
